\name{mines.custom}
\alias{mines.custom}
\title{Playing Mines}
\usage{
  mines.custom(n.row = 9, n.col = 9, n.mines = 16)
}
\arguments{
  \item{n.row}{The number of rows the minefield should
  contain}

  \item{n.col}{The number of columns the minefield should
  contain}

  \item{n.mines}{The number of mines in the minefield}
}
\description{
  Allows you to set the size of the board and the number of
  mines to be played with. A classic game that will provide
  entertainment to those in search of a challenge.
}
\author{
  Dason Kurkiewicz \email{dasonk@iastate.edu}
}

