\name{mines}
\alias{mines}
\title{A quick way to play Mines.}
\usage{
  mines(size = "small")
}
\arguments{
  \item{size}{The size of the minefield.  Can be either
  "small" "medium" or "large". Only the first letter
  matters.}
}
\value{
  Returns nothing but creates a game.
}
\description{
  This is an easier way to plays Mines.  Instead of using
  mines.custom to create the size of the board you want and
  to specify how many mines.  This allows you to use a few
  presets.
}
\author{
  Dason Kurkiewicz \email{dasonk@iastate.edu}
}

